refactor: 모달 컴포넌트와 스타일드 컴포넌트 분리
